Mod-settings fixed. Only the existing patches are shown. Each patch is not by default disabled instead of enabled. Removed references to removed patches.
