+++ Source/DocWorld.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29318.209
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32014.148
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DocWorld", "DocWorld\DocWorld.csproj", "{9D52D68C-2A24-4496-B70D-F7D275B9CB1E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DocWorldLite", "DocWorld\DocWorldLite.csproj", "{9D52D68C-2A24-4496-B70D-F7D275B9CB1E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
